client/requirements.txt
pyserial==3.1.1
requests==2.10.0
